Expose Oracles to allow access to imported enums
Allow access to the per-imported proto oracles to allow validation of
enum keys outside of the protobuf being parsed. Code which relies on
this library can only validate the enums parsed in the targeted
proto file. By exposing the enums found in the imported proto files we
can extend the validation to the set of all enums found during the
parse.